Decided to move forward with the generation stage for now
